+++ tests/Campaigns/CampaignsClientTest.php
@@ -414,6 +414,7 @@ JSON;
                     'headers' => [
                         'apiKey' => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent' => $this->userAgentHeader,
                     ],
                     'body'    => json_encode($payload),
                 ]
@@ -821,6 +822,7 @@ JSON;
                     'headers' => [
                         'apiKey' => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent' => $this->userAgentHeader,
                     ],
                     'body'    => $payload,
                 ]