+++ tests/Customers/CustomersClientTest.php
@@ -422,6 +422,7 @@ JSON;
                 'headers' => [
                     'apiKey' => '123456',
                     'Content-Type' => 'application/json',
+                    'User-Agent' => $this->userAgentHeader,
                 ],
                 'body'    => json_encode($payload),
             ])
@@ -584,6 +585,7 @@ JSON;
                 'headers' => [
                     'apiKey' => '123456',
                     'Content-Type' => 'application/json',
+                    'User-Agent' => $this->userAgentHeader,
                 ],
                 'body'    => json_encode($payload),
             ])