+++ tests/Licenses/LicensesClientTest.php
@@ -103,6 +103,7 @@ JSON;
                     'headers' => [
                         'apiKey'       => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent'   => $this->userAgentHeader,
                     ],
                     'body'    => json_encode($this->postData),
                 ]
@@ -189,6 +190,7 @@ JSON;
                     'headers' => [
                         'apiKey'       => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent'   => $this->userAgentHeader,
                     ],
                     'body'    => $expected,
                 ]
@@ -381,6 +383,7 @@ JSON;
                     'headers' => [
                         'apiKey'       => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent'   => $this->userAgentHeader,
                     ],
                     'body'    => json_encode($this->postData),
                 ]
@@ -623,6 +626,7 @@ JSON;
                     'headers' => [
                         'apiKey'       => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent'   => $this->userAgentHeader,
                     ],
                     'body'    => json_encode($this->postData),
                 ]
@@ -792,6 +796,7 @@ JSON;
                     'headers' => [
                         'apiKey'       => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent'   => $this->userAgentHeader,
                     ],
                     'body'    => json_encode($this->postData),
                 ]
@@ -849,6 +854,7 @@ JSON;
                     'headers' => [
                         'apiKey' => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent' => $this->userAgentHeader,
                     ],
                 ]
             )
@@ -869,6 +875,7 @@ JSON;
                     'headers' => [
                         'apiKey' => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent' => $this->userAgentHeader,
                     ],
                 ]
             )
@@ -918,6 +925,7 @@ JSON;
                     'headers' => [
                         'apiKey' => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent' => $this->userAgentHeader,
                     ],
                 ]
             )
@@ -961,6 +969,7 @@ JSON;
                     'headers' => [
                         'apiKey' => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent' => $this->userAgentHeader,
                     ],
                 ]
             )
@@ -1003,6 +1012,7 @@ JSON;
                     'headers' => [
                         'apiKey' => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent' => $this->userAgentHeader,
                     ],
                 ]
             )
@@ -1041,6 +1051,7 @@ JSON;
                     'headers' => [
                         'apiKey'       => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent'   => $this->userAgentHeader,
                     ],
                     'body'    => json_encode($postData),
                 ]
@@ -1087,6 +1098,7 @@ JSON;
                     'headers' => [
                         'apiKey' => '123456',
                         'Content-Type' => 'application/json',
+                        'User-Agent' => $this->userAgentHeader,
                     ],
                     'body'    => json_encode($postData),
                 ]