+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added a dedicated User-Agent header to the `PublicApiClient` requests 
+
 ## [0.11.0] - 2024-03-28
 
 - Added support of access tokens in all clients